\name{plfMA-package}
\alias{plfMA-package}
\docType{package}
\title{
A GUI to view, design and export various graphs of data
}
\description{
Provides a graphical user interface for viewing and designing various types of graphs of the data. A data objected can be created and loaded or a tab delimited file can be directly loaded in order to view the data as graphs. The graphs can be designed in different styles and exported in various formats of an image such as .jpeg, .png, .bmp, .pdf, and postscript.
}
\details{
	\tabular{ll}{
		Package: \tab plfMA\cr
		Type: \tab Package\cr
		Version: \tab 1.0.2\cr
		Date: \tab 2016-01-25\cr
		License: \tab GPL-2\cr
	}
	Just type pflMA() at R prompt to launch the GUI. Alternatively, create a data object such as x at R prompt and launch the GUI by typing plfMA(x).
}
\author{
Maintainer: Dhammapal Bharne <dhammapalb@uohyd.ac.in>
}
